pyproject.toml
[tool.poetry]
name = "paimon"
version = "0.1.0"
description = ""
authors = ["HUU <readabook123@protonmail.com>"]
packages = [
    { include = "paimon", from = "src" },
    { include = "beball", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^1.1.2"
watchdog = "^0.10.3"
requests = "^2.24.0"
python-daemon = "^2.2.4"
m3u8 = "^0.7.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
rope = "^0.18.0"
pyflakes = "^2.2.0"
pycodestyle = "^2.6.0"
pre-commit = "^2.7.1"
mypy = "^0.790"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
